+++ NEWS.md
@@ -1,6 +1,8 @@
 # tourr 0.6.1
 
-A radial tour is added that will allow a variable to be zero'd from a projection.
+* A radial tour is added that will allow a variable to be zero'd from a projection.
+
+* Fixes made according to CRAN requests
 
 # tourr 0.6.0